+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+### Changed
+- Changed OH Boost file to be climatology rather than single month
+
+### Added
+- Added a scaling factor to scale OH by a constant
+- Added exports for OH
+
 ## [1.0.0] - 2022-07-26
 
 ### Changed